slamd-96: Renamed algorithms/ -> experiment/, renamed ModelType -> ExperimentModel (+ filename change)
