+++ slamd/discovery/processing/experiment/experiment_model.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
 
-class ModelType(Enum):
+class ExperimentModel(Enum):
     RANDOM_FOREST = 'AI Model (lolo Random Forest)'
     GAUSSIAN_PROCESS = 'Statistics-based model (Gaussian Process Regression)'
